Fix connect-route dependency not resolving
At some point in the past, package.json contained a dependency to a personal fork of
connect-route that fixed a bug. The bug has since been fixed upstream and the fork
deleted, so "npm install" threw an error message stating that it couldn't find
the repository.

This commit changes the connect-route dependency to "*" so that it simply pulls in
the latest official version of connect-route.