Add CentOS sysv init file
This commit adds a sample sysv init file that can be used to manage
haste-server. It is tested to work properly on CentOS 6.

It assumes that haste-server is installed in /usr/local and that it's
supposed to be run by a dedicated 'haste' user account.

It can be adapted to suit users' needs.